pyproject.toml
[tool.poetry]
name = "300Tang"
version = "0.1.0"
description = ""
authors = ["yankihue <yankisirtas@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
